+++ neoforge/src/main/java/net/spell_engine/neoforge/compat/curios/SpellHostCurioItem.java
@@ -1,0 +1,37 @@
+package net.spell_engine.neoforge.compat.curios;
+
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.sound.SoundEvent;
+import top.theillusivec4.curios.api.SlotContext;
+import top.theillusivec4.curios.api.type.capability.ICurio;
+import top.theillusivec4.curios.api.type.capability.ICurioItem;
+
+public class SpellHostCurioItem extends Item implements ICurioItem {
+    private final SoundEvent equipSound;
+
+    public SpellHostCurioItem(Item.Settings settings, SoundEvent equipSound) {
+        super(settings);
+        this.equipSound = equipSound;
+    }
+
+    @Override
+    public boolean isEnchantable(ItemStack stack) {
+        return false;
+    }
+
+    @Override
+    public boolean canUnequip(SlotContext slotContext, ItemStack stack) {
+        var isOnCooldown = false;
+        if (slotContext.entity() instanceof PlayerEntity player) {
+            isOnCooldown = !player.isCreative() && player.getItemCooldownManager().isCoolingDown(stack.getItem());
+        }
+        return ICurioItem.super.canUnequip(slotContext, stack) && !isOnCooldown;
+    }
+
+    @Override
+    public ICurio.SoundInfo getEquipSound(SlotContext slotContext, ItemStack stack) {
+        return new ICurio.SoundInfo(this.equipSound, 1.0F, 1.0F);
+    }
+}